Setup empty home activity
