--- app/src/main/java/com/ensak/connect/ui/gallery/GalleryViewModel.java
@@ -1,19 +1,0 @@
-package com.ensak.connect.ui.gallery;
-
-import androidx.lifecycle.LiveData;
-import androidx.lifecycle.MutableLiveData;
-import androidx.lifecycle.ViewModel;
-
-public class GalleryViewModel extends ViewModel {
-
-    private final MutableLiveData<String> mText;
-
-    public GalleryViewModel() {
-        mText = new MutableLiveData<>();
-        mText.setValue("This is gallery fragment");
-    }
-
-    public LiveData<String> getText() {
-        return mText;
-    }
-}
