+++ app/src/main/res/navigation/mobile_navigation.xml
@@ -7,19 +7,7 @@
 
     <fragment
         android:id="@+id/nav_home"
-        android:name="com.ensak.connect.ui.home.HomeFragment"
+        android:name="com.ensak.connect.view.home.HomeFragment"
         android:label="@string/menu_home"
         tools:layout="@layout/fragment_home" />
-
-    <fragment
-        android:id="@+id/nav_gallery"
-        android:name="com.ensak.connect.ui.gallery.GalleryFragment"
-        android:label="@string/menu_gallery"
-        tools:layout="@layout/fragment_gallery" />
-
-    <fragment
-        android:id="@+id/nav_slideshow"
-        android:name="com.ensak.connect.ui.slideshow.SlideshowFragment"
-        android:label="@string/menu_slideshow"
-        tools:layout="@layout/fragment_slideshow" />
 </navigation>
